VQRSolution: Move to cvqf module
