+++ src/vqr/__init__.py
@@ -1,6 +1,6 @@
 # isort: skip_file
-from vqr.cvqf import QuantileFunction
-from vqr.solvers import VQRSolver, VQRSolution
+from vqr.cvqf import QuantileFunction, VQRSolution
+from vqr.solvers import VQRSolver
 from vqr.api import VectorQuantileEstimator, VectorQuantileRegressor
 from pkg_resources import DistributionNotFound, get_distribution
