+++ src/vqr/solvers/__init__.py
@@ -1,6 +1,7 @@
+from vqr.cvqf import VQRSolution
 from vqr.solvers.cvx import CVXVQRSolver
 from vqr.solvers.pot import POTVQESolver
-from vqr.solvers.base import VQRSolver, VQRSolution
+from vqr.solvers.base import VQRSolver
 from vqr.solvers.regularized_lse import (
     RegularizedDualVQRSolver,
     MLPRegularizedDualVQRSolver,
