+++ src/vqr/solvers/pot.py
@@ -4,10 +4,10 @@ import ot
 import numpy as np
 from numpy import ndarray as Array
 
-from vqr.cvqf import vector_quantile_levels
+from vqr.cvqf import VQRSolution, vector_quantile_levels
 from vqr.utils import get_kwargs
 from vqr.solvers.cvx import SIMILARITY_FN_INNER_PROD
-from vqr.solvers.base import VQRSolution, VQRDiscreteSolver
+from vqr.solvers.base import VQRDiscreteSolver
 
 
 class POTVQESolver(VQRDiscreteSolver):
